+++ components/train/src/DatasetLoader.py
@@ -109,7 +109,9 @@ class DatasetLoader(object):
                 for ij in range(index,index+self.batch_size):
                     # @note if there aren't enough independent
                     #       speakers, the batch won't be fillable
-                    spectrogram = numpy.load(self.data_list[ij][ii].replace(".wav", ".npy"))
+                    # @note casting pre-extracted float16 features to float32 so
+                    #       as to not inadvertently introduce network quantization
+                    spectrogram = numpy.load(self.data_list[ij][ii].replace(".wav", ".npy")).astype('float32')
                     spectrogram_torch_tensor = torch.FloatTensor(spectrogram)
                     feat.append(spectrogram_torch_tensor);
                 in_data.append(torch.cat(feat, dim=0));
